Move the new, gtk-related classes to their own files.
